refactor: Use permissions to check if user can view grid location (FLEX-1137)
